Move label services into Labels feature folders #388
