+++ code/FinanceManager.Api/Services/LabelSetterBackgroundService.cs
@@ -1,3 +1,4 @@
+using FinanceManager.Application.Labels.Setter;
 using FinanceManager.Application.Services.Ai;
 using FinanceManager.Domain.Entities.FinancialAccounts.Currencies;
 using FinanceManager.Domain.Repositories;
