+++ code/FinanceManager.Application/Labels/Suggestions/ILabelSuggestionPromptProvider.cs
@@ -1,4 +1,4 @@
-namespace FinanceManager.Application.Services.Ai;
+namespace FinanceManager.Application.Labels.Suggestions;
 
 public interface ILabelSuggestionPromptProvider
 {
