+++ code/FinanceManager.Infrastructure/Services/Ai/LabelSuggestionPromptProvider.cs
@@ -1,3 +1,4 @@
+using FinanceManager.Application.Labels.Suggestions;
 using FinanceManager.Application.Services.Ai;
 using System.Globalization;
